requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.0
gunicorn==21.2.0
sqlalchemy==2.0.40
pydantic==2.11.2
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.1.0
requests==2.31.0
python-linkedin-v2==0.9.4
slowapi==0.1.8
redis==5.0.1  # For rate limiting storage

# Development dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 